Move setting dates to its own module.
